move all logic to MovieListPage component
